requirements.txt
streamlit
openai
python-dotenv
schedule
pandas
numpy
requests